create seperate usepagefilter file
