+++ static/app/views/performance/browser/interactionsLandingPage.tsx
@@ -12,18 +12,15 @@ import PageFilterBar from 'sentry/components/organizations/pageFilterBar';
 import {ProjectPageFilter} from 'sentry/components/organizations/projectPageFilter';
 import {t} from 'sentry/locale';
 import {space} from 'sentry/styles/space';
-import {useDiscoverQuery} from 'sentry/utils/discover/discoverQuery';
-import EventView from 'sentry/utils/discover/eventView';
-import {DiscoverDatasets} from 'sentry/utils/discover/types';
 import {useLocation} from 'sentry/utils/useLocation';
 import useOrganization from 'sentry/utils/useOrganization';
-import usePageFilters from 'sentry/utils/usePageFilters';
 import {normalizeUrl} from 'sentry/utils/withDomainRequired';
 import InteractionsTable from 'sentry/views/performance/browser/interactionTable';
 import {
   BrowserStarfishFields,
   useBrowserModuleFilters,
 } from 'sentry/views/performance/browser/useBrowserFilters';
+import {usePagesQuery} from 'sentry/views/performance/browser/usePageQuery';
 import {ModulePageProviders} from 'sentry/views/performance/database/modulePageProviders';
 
 const {COMPONENT, PAGE, SPAN_ACTION} = BrowserStarfishFields;
@@ -141,38 +138,16 @@ function ActionSelector({value}: {value?: string}) {
 
 function PageSelector({value}: {value?: string}) {
   const location = useLocation();
-  const pageFilters = usePageFilters();
-  const {slug: orgSlug} = useOrganization();
 
-  const fields = ['transaction', 'p75(transaction.duration)', 'tpm()'];
-  const queryConditions = ['event.type:transaction', 'transaction.op:pageload'];
-
-  const eventView = EventView.fromNewQueryWithPageFilters(
-    {
-      fields, // for some reason we need a function, otherwise the query fails
-      name: 'Interaction module - page selector',
-      version: 2,
-      dataset: DiscoverDatasets.METRICS,
-      query: queryConditions.join(' '),
-      orderby: 'transaction',
-    },
-    pageFilters.selection
-  );
-
-  const {isLoading, data} = useDiscoverQuery({
-    eventView,
-    location,
-    orgSlug,
-    limit: 100,
-  });
+  const {data: pages, isLoading} = usePagesQuery();
 
   const options: Option[] =
-    !isLoading && data?.data.length
+    !isLoading && pages.length
       ? [
           {label: 'All', value: ''},
-          ...data?.data.map(row => ({
-            value: row.transaction.toString(),
-            label: row.transaction.toString(),
+          ...pages.map(page => ({
+            value: page,
+            label: page,
           })),
         ]
       : [];
