+++ static/app/views/performance/browser/useBrowserFilters.ts
